requirements.txt
moviepy==1.0.3
numpy==1.21.2
pydub==0.25.1
openai-whisper
